requirements.txt
colorama>=0.4,<1
pyyaml>=6.0,<7
charset-normalizer>=3.3,<4
watchdog>=4.0,<5
requests>=2.31,<3
rich>=13.7,<14
scikit-learn>=1.4,<1.6
joblib>=1.3,<2
pandas>=2.2,<3
matplotlib>=3.8,<4
scapy>=2.5,<3
pytest>=7.4,<8
